make string types bigger, redo alembic, other small changes
