+++ alembic/versions/561b31230723_add_lumen_twitter_tables.py
@@ -1,0 +1,199 @@
+"""add lumen twitter tables
+
+Revision ID: 561b31230723
+Revises: 16dbded8a5cf
+Create Date: 2017-04-17 02:28:10.599352
+
+"""
+
+# revision identifiers, used by Alembic.
+revision = '561b31230723'
+down_revision = '16dbded8a5cf'
+branch_labels = None
+depends_on = None
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import mysql
+
+def upgrade(engine_name):
+    globals()["upgrade_%s" % engine_name]()
+
+
+def downgrade(engine_name):
+    globals()["downgrade_%s" % engine_name]()
+
+
+
+
+
+def upgrade_development():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('lumen_notice_to_twitter_user',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('notice_id', sa.Integer(), nullable=True),
+    sa.Column('twitter_username', sa.String(length=256), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('lumen_notices',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('date_received', sa.DateTime(), nullable=True),
+    sa.Column('sender', sa.String(length=256), nullable=True),
+    sa.Column('principal', sa.String(length=256), nullable=True),
+    sa.Column('recipient', sa.String(length=256), nullable=True),
+    sa.Column('num_infringing_urls', sa.Integer(), nullable=True),
+    sa.Column('notice_data', mysql.MEDIUMTEXT(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('twitter_statuses',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('user_id', sa.BigInteger(), nullable=True),
+    sa.Column('in_reply_to_user_id', sa.BigInteger(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=True),
+    sa.Column('favorite_count', sa.Integer(), nullable=True),
+    sa.Column('retweet_count', sa.Integer(), nullable=True),
+    sa.Column('retweeted', sa.Boolean(), nullable=True),
+    sa.Column('status_data', mysql.MEDIUMTEXT(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('twitter_users',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('screen_name', sa.String(length=256), nullable=True),
+    sa.Column('name', sa.String(length=256), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=True),
+    sa.Column('followers_count', sa.Integer(), nullable=True),
+    sa.Column('friends_count', sa.Integer(), nullable=True),
+    sa.Column('lang', sa.String(length=64), nullable=True),
+    sa.Column('statuses_count', sa.Integer(), nullable=True),
+    sa.Column('verified', sa.Boolean(), nullable=True),
+    sa.Column('default_profile', sa.Boolean(), nullable=True),
+    sa.Column('default_profile_image', sa.Boolean(), nullable=True),
+    sa.Column('user_json', mysql.MEDIUMTEXT(), nullable=True),
+    sa.Column('user_state', sa.Integer(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade_development():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('twitter_users')
+    op.drop_table('twitter_statuses')
+    op.drop_table('lumen_notices')
+    op.drop_table('lumen_notice_to_twitter_user')
+    # ### end Alembic commands ###
+
+
+def upgrade_test():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('lumen_notice_to_twitter_user',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('notice_id', sa.Integer(), nullable=True),
+    sa.Column('twitter_username', sa.String(length=256), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('lumen_notices',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('date_received', sa.DateTime(), nullable=True),
+    sa.Column('sender', sa.String(length=256), nullable=True),
+    sa.Column('principal', sa.String(length=256), nullable=True),
+    sa.Column('recipient', sa.String(length=256), nullable=True),
+    sa.Column('num_infringing_urls', sa.Integer(), nullable=True),
+    sa.Column('notice_data', mysql.MEDIUMTEXT(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('twitter_statuses',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('user_id', sa.BigInteger(), nullable=True),
+    sa.Column('in_reply_to_user_id', sa.BigInteger(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=True),
+    sa.Column('favorite_count', sa.Integer(), nullable=True),
+    sa.Column('retweet_count', sa.Integer(), nullable=True),
+    sa.Column('retweeted', sa.Boolean(), nullable=True),
+    sa.Column('status_data', mysql.MEDIUMTEXT(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('twitter_users',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('screen_name', sa.String(length=256), nullable=True),
+    sa.Column('name', sa.String(length=256), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=True),
+    sa.Column('followers_count', sa.Integer(), nullable=True),
+    sa.Column('friends_count', sa.Integer(), nullable=True),
+    sa.Column('lang', sa.String(length=64), nullable=True),
+    sa.Column('statuses_count', sa.Integer(), nullable=True),
+    sa.Column('verified', sa.Boolean(), nullable=True),
+    sa.Column('default_profile', sa.Boolean(), nullable=True),
+    sa.Column('default_profile_image', sa.Boolean(), nullable=True),
+    sa.Column('user_json', mysql.MEDIUMTEXT(), nullable=True),
+    sa.Column('user_state', sa.Integer(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade_test():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('twitter_users')
+    op.drop_table('twitter_statuses')
+    op.drop_table('lumen_notices')
+    op.drop_table('lumen_notice_to_twitter_user')
+    # ### end Alembic commands ###
+
+
+def upgrade_production():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table('lumen_notice_to_twitter_user',
+    sa.Column('id', sa.Integer(), nullable=False),
+    sa.Column('notice_id', sa.Integer(), nullable=True),
+    sa.Column('twitter_username', sa.String(length=256), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('lumen_notices',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('date_received', sa.DateTime(), nullable=True),
+    sa.Column('sender', sa.String(length=256), nullable=True),
+    sa.Column('principal', sa.String(length=256), nullable=True),
+    sa.Column('recipient', sa.String(length=256), nullable=True),
+    sa.Column('num_infringing_urls', sa.Integer(), nullable=True),
+    sa.Column('notice_data', mysql.MEDIUMTEXT(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('twitter_statuses',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('user_id', sa.BigInteger(), nullable=True),
+    sa.Column('in_reply_to_user_id', sa.BigInteger(), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=True),
+    sa.Column('favorite_count', sa.Integer(), nullable=True),
+    sa.Column('retweet_count', sa.Integer(), nullable=True),
+    sa.Column('retweeted', sa.Boolean(), nullable=True),
+    sa.Column('status_data', mysql.MEDIUMTEXT(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    op.create_table('twitter_users',
+    sa.Column('id', sa.BigInteger(), nullable=False),
+    sa.Column('screen_name', sa.String(length=256), nullable=True),
+    sa.Column('name', sa.String(length=256), nullable=True),
+    sa.Column('created_at', sa.DateTime(), nullable=True),
+    sa.Column('followers_count', sa.Integer(), nullable=True),
+    sa.Column('friends_count', sa.Integer(), nullable=True),
+    sa.Column('lang', sa.String(length=64), nullable=True),
+    sa.Column('statuses_count', sa.Integer(), nullable=True),
+    sa.Column('verified', sa.Boolean(), nullable=True),
+    sa.Column('default_profile', sa.Boolean(), nullable=True),
+    sa.Column('default_profile_image', sa.Boolean(), nullable=True),
+    sa.Column('user_json', mysql.MEDIUMTEXT(), nullable=True),
+    sa.Column('user_state', sa.Integer(), nullable=True),
+    sa.PrimaryKeyConstraint('id')
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade_production():
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('twitter_users')
+    op.drop_table('twitter_statuses')
+    op.drop_table('lumen_notices')
+    op.drop_table('lumen_notice_to_twitter_user')
+    # ### end Alembic commands ###
+
