--- alembic/versions/8e190130ba36_add_twitter_user_table.py
@@ -1,109 +1,0 @@
-"""add twitter user table
-
-Revision ID: 8e190130ba36
-Revises: 9c8b30569bdf
-Create Date: 2017-04-16 16:21:40.214726
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '8e190130ba36'
-down_revision = '9c8b30569bdf'
-branch_labels = None
-depends_on = None
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-def upgrade(engine_name):
-    globals()["upgrade_%s" % engine_name]()
-
-
-def downgrade(engine_name):
-    globals()["downgrade_%s" % engine_name]()
-
-
-
-
-
-def upgrade_development():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('twitter_users',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('screen_name', sa.String(length=64), nullable=True),
-    sa.Column('name', sa.String(length=64), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('followers_count', sa.Integer(), nullable=True),
-    sa.Column('friends_count', sa.Integer(), nullable=True),
-    sa.Column('lang', sa.String(length=64), nullable=True),
-    sa.Column('statuses_count', sa.Integer(), nullable=True),
-    sa.Column('verified', sa.Boolean(), nullable=True),
-    sa.Column('default_profile', sa.Boolean(), nullable=True),
-    sa.Column('default_profile_image', sa.Boolean(), nullable=True),
-    sa.Column('user_json', mysql.MEDIUMTEXT(), nullable=True),
-    sa.Column('user_state', sa.Integer(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade_development():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('twitter_users')
-    # ### end Alembic commands ###
-
-
-def upgrade_test():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('twitter_users',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('screen_name', sa.String(length=64), nullable=True),
-    sa.Column('name', sa.String(length=64), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('followers_count', sa.Integer(), nullable=True),
-    sa.Column('friends_count', sa.Integer(), nullable=True),
-    sa.Column('lang', sa.String(length=64), nullable=True),
-    sa.Column('statuses_count', sa.Integer(), nullable=True),
-    sa.Column('verified', sa.Boolean(), nullable=True),
-    sa.Column('default_profile', sa.Boolean(), nullable=True),
-    sa.Column('default_profile_image', sa.Boolean(), nullable=True),
-    sa.Column('user_json', mysql.MEDIUMTEXT(), nullable=True),
-    sa.Column('user_state', sa.Integer(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade_test():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('twitter_users')
-    # ### end Alembic commands ###
-
-
-def upgrade_production():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('twitter_users',
-    sa.Column('id', sa.BigInteger(), nullable=False),
-    sa.Column('screen_name', sa.String(length=64), nullable=True),
-    sa.Column('name', sa.String(length=64), nullable=True),
-    sa.Column('created_at', sa.DateTime(), nullable=True),
-    sa.Column('followers_count', sa.Integer(), nullable=True),
-    sa.Column('friends_count', sa.Integer(), nullable=True),
-    sa.Column('lang', sa.String(length=64), nullable=True),
-    sa.Column('statuses_count', sa.Integer(), nullable=True),
-    sa.Column('verified', sa.Boolean(), nullable=True),
-    sa.Column('default_profile', sa.Boolean(), nullable=True),
-    sa.Column('default_profile_image', sa.Boolean(), nullable=True),
-    sa.Column('user_json', mysql.MEDIUMTEXT(), nullable=True),
-    sa.Column('user_state', sa.Integer(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade_production():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('twitter_users')
-    # ### end Alembic commands ###
-
