--- alembic/versions/9c8b30569bdf_add_lumen_tables.py
@@ -1,112 +1,0 @@
-"""add lumen tables
-
-Revision ID: 9c8b30569bdf
-Revises: 16dbded8a5cf
-Create Date: 2017-04-16 00:00:36.296975
-
-"""
-
-# revision identifiers, used by Alembic.
-revision = '9c8b30569bdf'
-down_revision = '16dbded8a5cf'
-branch_labels = None
-depends_on = None
-
-from alembic import op
-import sqlalchemy as sa
-from sqlalchemy.dialects import mysql
-
-def upgrade(engine_name):
-    globals()["upgrade_%s" % engine_name]()
-
-
-def downgrade(engine_name):
-    globals()["downgrade_%s" % engine_name]()
-
-
-
-
-
-def upgrade_development():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('lumen_notice_to_twitter_user',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('notice_id', sa.Integer(), nullable=True),
-    sa.Column('twitter_username', sa.String(length=64), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('lumen_notices',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('date_received', sa.DateTime(), nullable=True),
-    sa.Column('sender', sa.String(length=64), nullable=True),
-    sa.Column('principal', sa.String(length=64), nullable=True),
-    sa.Column('recipient', sa.String(length=64), nullable=True),
-    sa.Column('num_infringing_urls', sa.Integer(), nullable=True),
-    sa.Column('notice_data', mysql.MEDIUMTEXT(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade_development():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('lumen_notices')
-    op.drop_table('lumen_notice_to_twitter_user')
-    # ### end Alembic commands ###
-
-
-def upgrade_test():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('lumen_notice_to_twitter_user',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('notice_id', sa.Integer(), nullable=True),
-    sa.Column('twitter_username', sa.String(length=64), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('lumen_notices',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('date_received', sa.DateTime(), nullable=True),
-    sa.Column('sender', sa.String(length=64), nullable=True),
-    sa.Column('principal', sa.String(length=64), nullable=True),
-    sa.Column('recipient', sa.String(length=64), nullable=True),
-    sa.Column('num_infringing_urls', sa.Integer(), nullable=True),
-    sa.Column('notice_data', mysql.MEDIUMTEXT(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade_test():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('lumen_notices')
-    op.drop_table('lumen_notice_to_twitter_user')
-    # ### end Alembic commands ###
-
-
-def upgrade_production():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('lumen_notice_to_twitter_user',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('notice_id', sa.Integer(), nullable=True),
-    sa.Column('twitter_username', sa.String(length=64), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.create_table('lumen_notices',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('date_received', sa.DateTime(), nullable=True),
-    sa.Column('sender', sa.String(length=64), nullable=True),
-    sa.Column('principal', sa.String(length=64), nullable=True),
-    sa.Column('recipient', sa.String(length=64), nullable=True),
-    sa.Column('num_infringing_urls', sa.Integer(), nullable=True),
-    sa.Column('notice_data', mysql.MEDIUMTEXT(), nullable=True),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade_production():
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('lumen_notices')
-    op.drop_table('lumen_notice_to_twitter_user')
-    # ### end Alembic commands ###
-
